Fix typo for Hauppauge hybrid tuner recordings.
The linux driver for the analog side of the Hauppauge HVR-1600, cx18,
puts VBI data in Private Stream 1 in the MPEG stream. Each private stream
packet starts with "ITV0".

The linux driver for older analog Hauppauge tuners, ivtv, does the same,
except the private stream packets start with "itv0".

There was a typo in stream_functions.cpp that would have prevented it from
recognizing MPEG video captured by cx18 supported tuners. It was looking
for "VV0" when it should have been looking for "TV0".

It would make more sense to look for the full prefix strings,
"itv0" and "ITV0", but that would have required changing more than
a single character, and checking for a 3 character sequence has apparantly
been sufficient for PVR-x50 recordings.